Add files via upload
Establishing repo for this course. Got a new computer. Been struggling with an i3 and a core2 machines. Now have an i7. Anaconda installed in minutes, not hours. Hopeful. But gotta move all these files to the new environment. Seems like the perfect time to organize the mess of files I have into a coherent workspace.